remove ytdlp-node dependency, timeout cleanup

+++ src/download.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { incorrectUrl, getYtDlpPath, ytdlpConfig } from "./misc";
+import { incorrectUrl, getYtDlpPath, ytdlpConfig, timeouts } from "./misc";
 import { spawn } from "child_process";
 import fs from 'fs';
 import sanitize from "sanitize-filename";
@@ -19,7 +19,9 @@ router.post("/dl", async (req, res) => {
   stream.pipe(res);
 
   stream.on('close', () => {
-    fs.rmSync(dir, { recursive: true, force: true });
+    if (fs.existsSync(dir)) fs.rmSync(dir, { recursive: true, force: true });
+    clearTimeout(timeouts.get(`${sessionId}-${id}`));
+    timeouts.delete(`${sessionId}-${id}`);
   });
 
   stream.on('error', (err) => {
@@ -69,9 +71,12 @@ router.post('/', async (req, res) => {
       socket.emit("dl_ready", { message: { id } });
     });
 
-    setTimeout(() => {
-      if (fs.existsSync(outDir)) fs.rmSync(outDir, { recursive: true, force: true });
-    }, 60*60*1000);
+    timeouts.set(
+      `${sessionId}-${id}`,
+      setTimeout(() => {
+        if (fs.existsSync(outDir)) fs.rmSync(outDir, { recursive: true, force: true });
+      }, 60 * 60 * 1000)
+    );
   } catch (error) {
     res.status(500).send('Internal Server Error: ' + error);
   }
